requirements.txt
streamlit
pandas
numpy
sqlalchemy
kaggle
pymysql
patool
cryptography